+++ src/main/java/com/urassh/dvdrental/controller/rental/RentalCartController.java
@@ -5,6 +5,9 @@ import com.urassh.dvdrental.domain.Member;
 import com.urassh.dvdrental.domain.Money;
 import com.urassh.dvdrental.usecase.members.GetMemberUseCase;
 import com.urassh.dvdrental.usecase.rental.*;
+import com.urassh.dvdrental.usecase.rental.cart.ClearRentalCartUseCase;
+import com.urassh.dvdrental.usecase.rental.cart.GetRentalCartUseCase;
+import com.urassh.dvdrental.usecase.rental.cart.RemoveFromCartUseCase;
 import com.urassh.dvdrental.util.Navigator;
 import javafx.application.Platform;
 import javafx.beans.property.BooleanProperty;
