+++ src/main/java/com/urassh/dvdrental/usecase/rental/GetRentalsByMemberUseCase.java
@@ -1,4 +1,4 @@
-package com.urassh.dvdrental.usecase.goods;
+package com.urassh.dvdrental.usecase.rental;
 
 import com.urassh.dvdrental.domain.Member;
 import com.urassh.dvdrental.domain.Rental;
@@ -8,7 +8,7 @@ import com.urassh.dvdrental.infrastructure.RentalDummyRepository;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-public class GetRentedGoodsUseCase {
+public class GetRentalsByMemberUseCase {
     private final RentalRepository rentalRepository = new RentalDummyRepository();
 
     public CompletableFuture<List<Rental>> execute(Member member) {
