+++ src/main/java/com/urassh/dvdrental/usecase/rental/RemoveRentalUseCase.java
@@ -1,6 +1,5 @@
 package com.urassh.dvdrental.usecase.rental;
 
-import com.urassh.dvdrental.domain.Goods;
 import com.urassh.dvdrental.domain.Rental;
 import com.urassh.dvdrental.domain.interfaces.RentalRepository;
 import com.urassh.dvdrental.infrastructure.RentalDummyRepository;
