+++ src/main/java/com/urassh/dvdrental/usecase/rental/cart/GetRentalCountUseCase.java
@@ -1,4 +1,4 @@
-package com.urassh.dvdrental.usecase.rental;
+package com.urassh.dvdrental.usecase.rental.cart;
 
 import com.urassh.dvdrental.infrastructure.LocalStore;
