+++ src/main/java/com/urassh/dvdrental/usecase/rental/cart/RemoveFromCartUseCase.java
@@ -1,13 +1,13 @@
-package com.urassh.dvdrental.usecase.rental;
+package com.urassh.dvdrental.usecase.rental.cart;
 
 import com.urassh.dvdrental.domain.Goods;
 import com.urassh.dvdrental.infrastructure.LocalStore;
 
-public class AddToRentalCartUseCase {
+public class RemoveFromCartUseCase {
     private final LocalStore localStore = LocalStore.shared;
 
     public void execute(Goods goods) {
-        localStore.addToRentalCart(goods);
-        localStore.setRentalCount(localStore.getRentalCount() + 1);
+        localStore.removeFromRentalCart(goods);
+        localStore.setRentalCount(localStore.getRentalCount() - 1);
     }
 }
